refactored out key setup into ConsoleKeys. ConsoleReader is becoming too big and should be narrowed a bit in scope.
